requirements.txt
Pillow>=9.0.0
PyPDF2>=3.0.0
exifread>=3.0.0